[Quotes] Initial pass on i18n

+++ quotes/quotes.py
@@ -9,10 +9,14 @@ from odinair_libs.formatting import tick
 from odinair_libs.menus import ReactMenu, confirm, PostMenuAction, prompt
 
 from quotes.quote import Quote, conf
+from quotes.quote import i18n as _
 
 
 class Quotes:
     """Save and retrieve quotes"""
+
+    __author__ = "odinair <odinair@odinair.xyz>"
+    __version__ = "0.1.0"
 
     def __init__(self, bot: Red):
         self.bot = bot
@@ -24,7 +28,7 @@ class Quotes:
         """Save and retrieve quotes"""
         quote = await Quote.get(ctx.guild, quote)
         if quote is None:
-            await ctx.send(warning("That quote doesn't exist"))
+            await ctx.send(warning(_("That quote doesn't exist")))
             return
         await ctx.send(embed=quote.embed)
 
@@ -32,7 +36,7 @@ class Quotes:
     async def _quote_add(self, ctx: RedContext, *, message: str):
         """Add a quote"""
         quote = await Quote.create(message, ctx.author, ctx.author)
-        await ctx.send(tick("Quote added"), embed=quote.embed)
+        await ctx.send(tick(_("Quote added")), embed=quote.embed)
 
     @quote.command(name="message")
     async def _quote_message(self, ctx: RedContext, message: int):
@@ -46,12 +50,12 @@ class Quotes:
         try:
             message = await ctx.get_message(message)
         except discord.NotFound:
-            await ctx.send(warning("I couldn't find that message. (is it in a different channel?)"))
+            await ctx.send(warning(_("I couldn't find that message. (is it in a different channel?)")))
         except discord.Forbidden:
-            await ctx.send(warning("I'm not allowed to retrieve that message"))
+            await ctx.send(warning(_("I'm not allowed to retrieve that message")))
         else:
             quote = await Quote.create(message.content, ctx.author, message.author)
-            await ctx.send(tick("Quote added"), embed=quote.embed)
+            await ctx.send(tick(_("Quote added")), embed=quote.embed)
 
     @quote.command(name="edit", aliases=["modify"])
     async def _quote_edit(self, ctx: RedContext, quote: int):
@@ -62,18 +66,18 @@ class Quotes:
         """
         quote = await Quote.get(ctx.guild, quote)
         if quote is None:
-            await ctx.send(warning("That quote doesn't exist"))
+            await ctx.send(warning(_("That quote doesn't exist")))
             return
         if not await quote.can_modify(ctx.author):
-            await ctx.send(warning("You aren't authorized to modify that quote"))
+            await ctx.send(warning(_("You aren't authorized to modify that quote")))
             return
 
-        embed = discord.Embed(title="Edit Quote",
-                              description="What action(s) would you like to take?\n\n"
-                                          "\N{BUST IN SILHOUETTE} \N{EM DASH} Attribute quote\n"
-                                          "\N{MEMO} \N{EM DASH} Edit content\n"
-                                          "\N{WASTEBASKET} \N{EM DASH} Delete quote\n"
-                                          "\N{CROSS MARK} \N{EM DASH} Cancel")
+        embed = discord.Embed(title=_("Edit Quote"),
+                              description=_("What action(s) would you like to take?\n\n"
+                                            "\N{BUST IN SILHOUETTE} \N{EM DASH} Attribute quote\n"
+                                            "\N{MEMO} \N{EM DASH} Edit content\n"
+                                            "\N{WASTEBASKET} \N{EM DASH} Delete quote\n"
+                                            "\N{CROSS MARK} \N{EM DASH} Cancel"))
 
         actions = {
             "attribute": "\N{BUST IN SILHOUETTE}",
@@ -89,7 +93,7 @@ class Quotes:
                     break
 
                 elif result == "attribute":
-                    msg = await prompt(ctx, content="Who would you like to attribute this quote to?", timeout=30.0,
+                    msg = await prompt(ctx, content=_("Who would you like to attribute this quote to?"), timeout=30.0,
                                        delete_messages=True)
                     if not msg:
                         continue
@@ -97,17 +101,18 @@ class Quotes:
                     try:
                         member = await commands.MemberConverter().convert(ctx, msg.content)
                     except commands.BadArgument:
-                        await ctx.send(warning(f"Failed to convert `{msg.content}` into a Member - try again?"),
+                        await ctx.send(warning(_("Failed to convert `{}` into a Member - try again?")
+                                               .format(msg.content)),
                                        delete_after=30.0)
                         continue
                     else:
                         quote.edited = True
                         quote.message_author = member
                         await quote.save()
-                        await ctx.send(tick(f"Attributed quote to **{member!s}**."), delete_after=30.0)
+                        await ctx.send(tick(_("Attributed quote to **{}**.").format(str(member))), delete_after=30.0)
 
                 elif result == "edit_content":
-                    msg = await prompt(ctx, content="Please enter the new content for the quote", timeout=120.0,
+                    msg = await prompt(ctx, content=_("Please enter the new content for the quote"), timeout=120.0,
                                        delete_messages=True)
                     if not msg:
                         continue
@@ -115,14 +120,14 @@ class Quotes:
                     quote.edited = True
                     quote.text = msg.content
                     await quote.save()
-                    await ctx.send(tick("Modified quote contents successfully."), delete_after=30.0)
+                    await ctx.send(tick(_("Modified quote contents successfully.")), delete_after=30.0)
 
                 elif result == "delete":
-                    if await confirm(ctx, message="Are you sure you want to delete this quote?\n\n"
-                                                  "**This action cannot be undone!**",
+                    if await confirm(ctx, message=_("Are you sure you want to delete this quote?\n\n"
+                                                    "**This action cannot be undone!**"),
                                      colour=discord.Colour.red()):
                         await quote.delete()
-                        await ctx.send(tick("Quote deleted."), delete_after=30.0)
+                        await ctx.send(tick(_("Quote successfully deleted.")), delete_after=30.0)
                         break
 
         try:
@@ -138,16 +143,16 @@ class Quotes:
         """
         quote = await Quote.get(ctx.guild, quote)
         if quote is None:
-            await ctx.send(warning("That quote doesn't exist"))
+            await ctx.send(warning(_("That quote doesn't exist")))
             return
         if not await quote.can_modify(ctx.author):
-            await ctx.send(warning("You aren't allowed to modify that quote"))
+            await ctx.send(warning(_("You aren't authorized to modify that quote")))
             return
 
         quote.edited = True
         quote.message_author = author
         await quote.save()
-        await ctx.send(tick(f"Attributed quote #{quote} to **{author!s}**."))
+        await ctx.send(tick(_("Attributed quote #{} to **{}**.").format(quote, str(author))))
 
     @quote.command(name="remove", aliases=["rm", "delete"])
     async def _quote_remove(self, ctx: RedContext, quote: int):
@@ -157,11 +162,16 @@ class Quotes:
         """
         quote = await Quote.get(ctx.guild, quote)
         if not quote:
-            await ctx.send(warning("That quote doesn't exist"))
+            await ctx.send(warning(_("That quote doesn't exist")))
             return
         if not await quote.can_modify(ctx.author):
-            await ctx.send(warning("You aren't authorized to remove that quote"))
+            await ctx.send(warning(_("You aren't authorized to remove that quote")))
             return
 
-        await quote.delete()
-        await ctx.send(tick("Quote removed."))
+        if await confirm(ctx, message=_("Are you sure you want to delete this quote?\n\n"
+                                        "**This action cannot be undone!**"),
+                         colour=discord.Colour.red()):
+            await quote.delete()
+            await ctx.send(tick(_("Quote successfully deleted.")))
+        else:
+            await ctx.send(_("Ok then."))
